+++ src/components/EzHeading/index.js
@@ -1,0 +1,3 @@
+import EzHeading from './EzHeading';
+
+export default EzHeading;