+++ src/index.js
@@ -3,6 +3,7 @@ import * as themes from './themes';
 export {default as EzAlert} from './components/EzAlert';
 export {default as EzButton} from './components/EzButton';
 export {EzCard, EzCardSection} from './components/EzCard';
+export {default as EzHeading} from './components/EzHeading';
 export {default as EzLabelledItem} from './components/EzLabelledItem';
 export {EzPageContent, EzContentGroup} from './components/EzPageContent';
 export {default as EzSegmentedControl} from './components/EzSegmentedControl';